Main page content

+++ FitnessPortalAPI/Entities/Exercise.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public int NumberOfReps { get; set; }
-        public int Payload { get; set; }
+        public float Payload { get; set; }
         public int TrainingId { get; set; }
         public virtual Training Training { get; set; }
     }
